Refactor: carousel을 동적함수와 contentWrapper로 분리
